rename ScalingTaskScheduler to JobScheduler
